feat: trim activity screens to error logs only in user app
