+++ backend/src/homepot/app/api/API_v1/Endpoints/__init__.py
@@ -1,7 +1,6 @@
 """Endpoints package for API v1."""
 
 from . import (
-    AgentActivityEndpoint,
     AgentAlertEndpoint,
     AgentAuditEndpoint,
     AgentConfigHistoryEndpoint,
@@ -35,7 +34,6 @@ from . import (
 )
 
 __all__ = [
-    "AgentActivityEndpoint",
     "AgentAlertEndpoint",
     "AgentAuditEndpoint",
     "AgentConfigHistoryEndpoint",
